chore: convert to esm
